[ADD] 대결 메인 화면 대기 중인 대결 처리 API 구현 완료

+++ domain/src/main/java/kr/co/domain/model/WaitingBattle.kt
@@ -1,12 +1,12 @@
 package kr.co.domain.model
 
 data class WaitingBattle(
-    val battleId: Int,
+    val battleId: Long,
     val duration: Int,
-    val otherId: Int,
+    val otherId: Long,
     val otherNickname: String,
     val otherProfileImg: String,
     val status: String,
     val word: String,
-    val wordId: Int
+    val wordId: Long
 )
